requirements.txt
acvl_utils==0.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
batchgenerators==0.21
beautifulsoup4==4.13.3
bleach==6.1.0
certifi==2025.1.31
charset-normalizer==2.0.12
connected-components-3d==3.22.0
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dicom2nifti==2.6.0
dill==0.4.0
docopt==0.6.2
dynamic_network_architectures==0.3.1
einops==0.8.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
fury==0.12.0
future==1.0.0
h5py==3.6.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.9
linecache2==1.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
MedPy==0.5.2
mistune==3.1.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
nibabel==5.2.0
nnunet-customized==1.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
p_tqdm==1.4.2
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydicom==2.4.4
pygltflib==1.16.4
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-gdcm==3.0.25
pytorch-lightning==2.4.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.35.1
requests==2.27.1
rpds-py==0.20.1
rt-utils==1.2.7
scikit-image==0.19.2
scikit-learn==1.3.2
scipy==1.10.0
seaborn==0.13.2
SimpleITK==2.3.1
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tensorly==0.9.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.2.1
torch==2.4.1
torchmetrics==1.4.2
torchsummary==1.5.1
torchvision==0.19.1
tornado==6.4.2
TotalSegmentator==1.5.7
tqdm==4.66.2
traceback2==1.4.0
traitlets==5.14.3
triton==3.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
urllib3==1.26.20
vtk==9.3.1
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
xvfbwrapper==0.2.10
yacs==0.1.8
yarg==0.1.9
yarl==1.15.2
zipp==3.20.2
